Move shell tests into separate test file
